Rename file for getting main programming language (typing error) #2710
